Swap class names RowBuilder <> Row
